Use -retry-join to avoid cluster bootstrap race conditions
If a non-bootstrap node starts up, finds another node in etcd, but that
node hasn't yet gotten consul up and listening for joins, then the
non-bootstrap node wil fail to join and quit.  Consul provides the
-retry-join flag that will retry joining the host a number of times at a
configured interval.  So, use this but let the consumer of this
container set the number of retries and interval.