Uses etcd REST API, no longer needs etcdctl
Removes the need for etcdctl to be mapped into the container via a
volume.  Also uses the quorum=true flag when getting values to attempt
to avoid bootstrap race conditions.  Finally, fixes an issue where we
would add a --retry-join flag for Consul with no value (which causes the
startup to fail).